Issue 13124: Merged changes.
git-svn-id: https://svn.imcode.com/imcode/imcms/branches/6.0.0-a52-imagearchive-i13124@30563 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/imcode/util/image/Format.java
@@ -16,10 +16,10 @@ public enum Format {
     PDF("PDF", 10, "pdf", "application/pdf", true),
     PS("PS", 11, "ps", "application/postscript", true);
 
-    
+
     private static final Map<String, Format> FORMAT_MAP = new HashMap<String, Format>(Format.values().length);
     private static final Map<String, Format> EXTENSION_MAP = new HashMap<String, Format>(Format.values().length);
-    private static final Map<Short, Format> ORDINAL_MAP = new HashMap<Short, Format>(Format.values().length);
+    private static final Map<Integer, Format> ORDINAL_MAP = new HashMap<Integer, Format>(Format.values().length);
     static {
         for (Format format : Format.values()) {
             FORMAT_MAP.put(format.getFormat(), format);
@@ -27,28 +27,28 @@ public enum Format {
             ORDINAL_MAP.put(format.getOrdinal(), format);
         }
     }
-    
+
     private final String format;
-    private final short ordinal;
+    private final int ordinal;
     private final String extension;
     private final String mimeType;
     private final boolean writable;
 
-    
+
     private Format(String format, int ordinal, String extension, String mimeType, boolean writable) {
         this.format = format;
-        this.ordinal = (short) ordinal;
+        this.ordinal = ordinal;
         this.extension = extension;
         this.mimeType = mimeType;
         this.writable = writable;
     }
 
-    
+
     public String getFormat() {
         return format;
     }
 
-    public short getOrdinal() {
+    public int getOrdinal() {
         return ordinal;
     }
 
@@ -63,17 +63,17 @@ public enum Format {
     public String getMimeType() {
         return mimeType;
     }
-    
-    
+
+
     public static Format findFormat(String format) {
         return FORMAT_MAP.get(format);
     }
-    
+
     public static Format findFormatByExtension(String extension) {
-    	return EXTENSION_MAP.get(extension);
+        return EXTENSION_MAP.get(extension);
     }
-    
-    public static Format findFormat(short ordinal) {
-    	return ORDINAL_MAP.get(ordinal);
+
+    public static Format findFormat(int ordinal) {
+        return ORDINAL_MAP.get(ordinal);
     }
 }
